+++ tests/integration/annotation_import/conftest.py
@@ -361,6 +361,19 @@ def model_run(rand_gen, model):
 
 
 @pytest.fixture
+def model_run_with_training_metadata(rand_gen, model):
+    name = rand_gen(str)
+    training_metadata = {"batch_size": 1000}
+    model_run = model.create_model_run(name, training_metadata)
+    yield model_run
+    try:
+        model_run.delete()
+    except:
+        # Already was deleted by the test
+        pass
+
+
+@pytest.fixture
 def model_run_with_model_run_data_rows(client, configured_project,
                                        model_run_predictions, model_run):
     configured_project.enable_model_assisted_labeling()